uefi: Rework `exit_boot_services` API
Exiting boot services is a one-way operation, which is why `exit_boot_services`
takes `self` rather than a reference. If it succeeds then you get back a
`SystemTable<Runtime>`, but even if it fails you can't continue to use boot
services except for a very limited subset. Specifically, in UEFI 2.8 and
earlier, only `get_memory_map` and `exit_boot_services` are allowed. Starting in
UEFI 2.9 other memory allocation functions may also be called.

To properly support pre-2.9 firmware then, the right thing to do is to
make sure that the initial buffer allocated to hold the memory map has
sufficient extra room in case the first attempt to get the memory map
and exit boot services fails. We now add room for 8 extra entries if
that occurs, matching the behavior of the Linux kernel:
https://github.com/torvalds/linux/blob/e544a07438/drivers/firmware/efi/libstub/efistub.h#L173

If something unexpected happens and exiting boot services still fails
after two attempts, the system is in an undefined state that we can't
reasonably recover from. The best thing we can do then is to simply
reset the system, essentially a strong panic.