wip: Change toolchain to `beta`
This is a WIP commit for switching to the stable toolchain when Rust 1.68 is
released (scheduled for March 9). For purposes of testing, the toolchain is set
to `beta` instead of `stable` for now.

In `msrv_toolchain.toml`, the channel will be changed to `1.68`.

Some of the uefi-macros compiler ui tests have been modified so that the error
output matches the errors produced by stable compiler, which are currently
slightly different in a few cases from the nightly compiler. In particular, the
precise spans we output for errors in the guid macros don't work on stable,
because proc-macro's Literal::subspan method is currently unstable.