test-runner: Make `unstable` an optional feature
Add an `unstable` option to `uefi-test-runner` that controls whether the
`unstable` feature of the `uefi` crate is enabled. Also add a separate CI job so
that we test both with and without the unstable feature.

This change is a necessary prereq for testing on stable, since we can't
unconditionally enable unstable features there.