ci: Add a nightly_channel job
This job is in preparation for switching most of the CI to operating on the
stable channel. We still need some actions to run on the nightly channel so we
can test the `unstable` feature and run tests under miri.

To that end, move various steps under a single `nightly_channel` job. The job
switches to the nightly channel using a new `nightly_toolchain.toml` file, so
once we switch the top-level `rust-toolchain.toml` to stable, the new job will
remain on nightly.