xtask: Add option to skip uefi-macros tests
The uefi-macros tests check that the expected compiler output is produced for
errors. The error output is slightly different between stable and nightly, so
when we transition to testing on stable we have to handle that somehow. We could
potentially have two sets of expected output, one for stable and one for
nightly, but that's probably more trouble than it's worth. Instead, add an
option to skip the uefi-macros tests, and enable that flag in the CI job for the
nightly channel.