uefi-macros: Make `entry` example more compatible with stable
The `entry` macro's example code had `#![no_std]` in it to make it look like how
you'd normally use it with the uefi targets. In the context that the example
code is compiled in though, the target is the host's target. To get the example
compiling there, we had to add some hidden hacks for `eh_personality` and a
panic handler. Those hacks don't work on the stable channel though, so drop them
along with `no_std`.