Rework "Building UEFI programs" sections in the readmes
These sections were mostly duplicated between the top-level readme and the
`uefi` package's readme. In the top-level readme, it had a broken link to the
`BUILDING.md` document which has been deleted.

Since we now have "Documentation" sections in both those readmes, the "Building
UEFI programs" section was mostly redundent, except for the part about MSRV. Add
new MSRV sections, and in the top-level readme just link to the `uefi` package's
readme so that we don't have to update both places when the MSRV changes.