uefi: Drop `'boot` lifetime from Output protocol
The `Output` protocol has an internal pointer to an `OutputData` object. This
was represented as a reference with the `'boot` lifetime, but this isn't quite
right since the actual lifetime could be shorter than boot
services. Furthermore, in several places the lifetime parameter was set to
`'static`, which is definitely not correct.

Fix by dropping the lifetime parameter, and instead just use a raw pointer. This
is more in line with how we've implemented other protocols.