add Cargo.lock
This will increase our CI stability, as we also use the lock file for some binaries.
However, as the lock file is on the top-level of the workspace, it will not be packaged
into library releases. This is also recommended for libraries by the ecosystem.

Additionally, this will enable us to use better CI caching to accelerate our CI.