ci: Drop test_latest_release job
This job was useful when we were on the nightly channel and using various
unstable features that could break at any time, but there's not much value to
keeping it now that we're on the stable channel.

The other minor thing the job checked was that the specific build command
provided in the tutorial (originally in `BUILDING.md`) works. That was helpful
back when we had some unstable (and verbose) `-Zbuild-std` flags, but now the
build command is just `cargo build --target x86_64-unknown-uefi` which doesn't
really need testing.