uefi: Move `Status::into_with*` methods to a new `StatusExt` trait
This is in preparation for moving `Status` to uefi-raw. since the `Error` type
and `Result` type alias will not move to uefi-raw, we can't move the `into_with`
methods to uefi-raw. And since we can't impl methods on a type defined in
another crate, we have to move these methods to a new trait.