Use the uguid crate to replace the `Guid` struct and `guid!` macro
This should be mostly transparent for the end user except that the `from_values`
method has been removed. Usually the `guid!` macro is a more convenient choice
anyway, but `new` or `from_bytes` can also be used if needed.

This allows us to delete a bit of code and some tests, as well as slightly
reducing our dependency on proc macros as the `uguid::guid!` macro is a
declarative macro.