uefi-raw: Add table Header struct
This slightly differs from the version in the `uefi` crate in that the
`reserved` field is public (to make it easy for callers to construct a `Header`,
something that users of `uefi` would not need to do normally). Since this is a
small structure, I didn't bother trying to share code between the two.