Make UEFI shell protocols testable
Add a new `shell_launcher` app that gets booted first when launching QEMU. This
app launches the shell app (which is built alongside the OVMF files). The shell
app's command line is set so that it will launch the test-runner app. Doing it
this way allows us to avoid the shell app's built-in five second startup delay.

By launching the test-runner inside the shell, in future commits we'll be able
to test UEFI shell protocols.