xtask: Improve QEMU boot time
On aarch64 it sits for a while on the initial screen and shows a progress bar to
give the user time to enter the boot menu. We don't want to do that in our
tests, so set the timeout to zero.

Even though we're skipping the boot menu, we have to also enable it here,
otherwise the timeout value is ignored and a hardcoded platform default of 3
seconds is used.