xtask: Add check-raw
Running `cargo xtask check-raw` validates various properties of the code in the
`uefi-raw` package. For example, it checks that no Rust enums are used, that
structs have an appropriate repr for FFI, that raw pointers are used instead of
references, and many other things.

This is run in CI as part of the lint job.