uefi: Simplify install_protocol_interface() as friends
We can use the type system to get at the protocol GUID. More
importantly, this makes sure to mention that the installed instances
must stay alive for the duration that they are installed.

Also mention that null pointers are allowed, but must only be used on
protocols intetended to be used as markers.