uefi: Remove some uses of MaybeUninit in BootServices
While `MaybeUninit` is useful to avoid unnecessarily initializing large buffers,
it's not needed for small values like single pointers and integers.

For `Event`, `Handle`, and `ProtocolSearchKey`, which internally contain a
`NonNull` pointer, we can wrap them in an `Option` and still have the same
layout as a raw pointer. The initial value can then be `None` rather than
`MaybeUninit`.

This has the nice side effect of removing a little bit of `unsafe` code.