uefi: Improve support for null protocol interfaces
Modify `ScopedProtocol` to check if the interface pointer is null. If it is,
then `Deref` and `DerefMut` will now panic. They are also now marked
`#[track_caller]` so that the panic location points to the caller.

Also added `get` and `get_mut` methods to get potentially-null interface data
without panicking.