ShellParams: Add subshell test
To test whether the ShellParams protocol works properly, we should try
it with different parameters, not just the default of one.

To do that we need to run the test runner again. So that it doesn't run
all of the tests again, we test whether the ShellParams protocol is
installed and whether arguments have been passed. If it's not installed
or if there are no arguments, the default tests run.
As part of ShellParams test, it launches the second instance of
test_runner.efi  with parameters. That instance will only run a specific
subshell test to check if the parameters were correctly passed.

Signed-off-by: Daniel Schaefer <dhs@frame.work>