Use const interface pointers in protocol management functions
A protocol's interface may mutate itself, but the void pointer passed to the
`{install,reinstall,uninstall}_protocol_interface` functions is opaque to the
firmware, so there's no need for it to be a mut pointer.

Note that from a Rust safety perspective there's no difference here -- mut and
const pointers are interchangeable.

https://github.com/rust-osdev/uefi-rs/issues/970