uefi-raw: Fill in [un]install_multiple_protocol_interfaces pointers
The existing comment there said the unstable `c-variadic` feature was required,
but that was incorrect. While defining c-variadic functions is unstable in Rust,
c-variadic function pointers are stable (calling c-variadics is also stable).

We still can't quite define these two function pointers correctly because they
should be `extern "efiapi", which requires the unstable
`extended_varargs_abi_support` feature. However, as long as your code is
compiled for a UEFI target, `extern "C"` gives you the correct ABI. Since these
function pointers are `unsafe` to call, we can expose them in the raw interface
and leave it up to the caller to ensure they are used correctly.