fs: Remove a couple debug logs
The trace log in `open` can be a bit misleading since at first glance it looks
like an error, but in some cases it's expected for open to fail, like when
calling `try_exists`.

The debug log in `create_dir_all` is showing internal details so probably not
all that useful to end users.