uefi-test-runner: Fix missing log messages
After opening the serial protocol in exclusive mode we have to reconnect it to
the output console or we'll drop logs. After the recent upgrade to edk2 this
code stopped working. Fix it by specifying the terminal type when calling
`connect_controller`.

This is needed because of edk2 commit cf6a0a52b07: "use utf8 for the serial
console". The default terminal type (for OVMF) was changed from PC-ANSI to
UTF-8. However, in TerminalDriverBindingStart, the terminal type will be reset
to the default value of PC-ANSI if `remaining_device_path` is null. That in turn
will prevent the code from finding the VT-UTF8 device.

The terminal type on aarch64 is VT 100, so switch based on the target aarch.