uefi: Drop the panic-on-logger-errors feature
This was originally added because of occasional panics when logging quickly on
VirtualBox. (Unclear whether this bug is still present, and also as far as I
know we haven't observed this behavior on any other UEFI implementations.) The
decision was made to panic by default on logger errors, but offer an escape
mechanism. However, making this a compile-time choice is not ideal, since most
UEFI programs are intended to run on arbitrary UEFI implementations.

We could make it a runtime option instead, but since loggers are usually just
informational (i.e. not critical functionality for the application), silently
ignoring errors seems like a better choice for most uses.

In the rare case where an application does consider logging critical, they can
turn off the `logger` helper and implement their own logger.

For prior discussion, see:
* https://github.com/rust-osdev/uefi-rs/issues/121
* https://github.com/rust-osdev/uefi-rs/issues/123