Fix uefi-macros trybuild test
A difference in behavior between CI and local testing was observed for the
entry_bad_return_type test. This is due to the way trybuild sets the `--verbose`
flag, which is overwritten if RUSTFLAGS is set (as it always is for our CI
jobs). Add code in xtask to ensure that `--verbose` still gets set correctly.

Also update entry_bad_return_type.stderr to match the expected current output.