uefi: introduce types MemoryMapOwned, MemoryMapRef, and MemoryMapRefMut
This provides API users all the flexibility they need. The main motivation for
the addition is that one can use a chunk of memory in a kernel (provided by a
bootloader) and parse it as EFI memory map. The main motivation for the specific
implementation (via traits) is code reduction.