runtime: Add variable key functions
The interface is slightly different than in `uefi/src/table/runtime.rs`,
where we just have a function to get all keys at once in a `Vec<VariableKey>`.

For the new implementation, two interfaces are
provided:

`get_next_variable_key` is a low-level interface that does not require
the `alloc` feature; it's a simple wrapper around `GetNextVariableName`.

`variable_keys` is a high-level interface that returns an
iterator. (Following the conventions of Rust's std lib, the iterator
type is named `VariableKeys`.) This requires `alloc`.