uefi: mem: add public constructors for MemoryMapRef and MemoryMapRefMut
There is no public constructor for MemoryMapOwned, as the construction of it is
quite complicated and therefore hidden behind BootServices::memory_map() respectively
SystemTable::exit_boot_services.

The real-world use case for the public constructors for the other types is when
parsing the UEFI memory map from kernels after a bootloader, such as GRUB or Limine,
passed the UEFI memory map via the corresponding boot information.