clippy/uefi: prefer const fn where possible
This has the benefit that users clearly see that no side effects can happen.
However, the public API changes in the sense that once we remove const for a
function, user code might break.

Nevertheless, this is a code improvement, IMHO.

Regarding the device path code, which is generated: So far the simplest solution
is to deactivate this lint there. It is not trivial to decide which function can
be const and which not during code generation. It is future work to solve this,
if necessary.