doc: streamline sections Contributing, MSRV, and License
Copy MSRV, License, and Contributing sections to lib.rs and streamline it
with the README equivalents. We should also include these information in
lib.rs, as it is the main entry point into everything relevant about the
library. At least, it should from now on.

The big benefit is that the documentation on `docs.rs` then covers
everything relevant.