boot: Add freestanding stall
One minor implementation change from the BootServices version: the Status value
is ignored rather than asserting it is `SUCCESS`. There shouldn't be any
practical difference, as stalling should always succeed. In the unlikely event
that some implementation returns an error, a panic is unlikely to be desirable
to the caller.