ci: cancel previous runs on further pushes
The idea here is to cancel all runs that are in progress, when a new push to a branch
happens. Typically, a pipeline is invalid as soon as changes are merged. A new pipeline
will start anyway.

To save computing time, and to prevent users from many expected "pipeline has failed"
messages, this is beneficial.

The impact is especially visible if you do frequent pushes in a short period of time.

More info: https://stackoverflow.com/questions/66335225/how-to-cancel-previous-runs-in-the-pr-when-you-push-new-commitsupdate-the-curre