uefi-macros: Rename generated entry arguments
When the `entry` macro is used on a no-argument function, system table and
handle arguments are generated so that the table/handle globals can be
initialized. These arguments are not meant for other use. (There's no technical
issue with the body of the user-supplied function using these arguments, but it
makes the code confusing since the arguments appear out of nowhere, and it's a
compatibility hazard since the name and type of the arguments may change in the
future.)

To make this less likely to occur, prepend `internal_` to the arg names.