uefi: Fix compilation of minimal example
This turns out to be easier than I expected: just add `extern crate std;` so
that the panic handler and global allocator are set. The line is hidden so that
it doesn't appear in the rendered docs.

This allows the code block to `no_run` instead of `ignore`, so it will be
compiled (but not run, so there's no `main` linker error), and that also has the
nice side effect of removing the orange `!` warning bubble in the rendered docs.