Revert "ci: check if crate can be packaged"
This reverts commit fc5c0a98c7f19acb9ccada14cd6a63d4c85e9b62.

When `cargo package` builds `uefi`, it does so with dependencies from
crates.io. That causes a problem if `uefi` depends on some not-yet-released API
changes in `uefi-raw` or `uefi-macros`. For example, see:
https://github.com/rust-osdev/uefi-rs/pull/1383#issuecomment-2315661178

I couldn't find a way to force it to use local deps, so revert the job for now.