xtask: Update OVMF prebuilts
edk2-stable202402 is the latest release that passes CI. Starting in
edk2-stable202405, the aarch64 job fails early in boot with a "Synchronous
Exception" error.

This seems likely to be an issue with the version of QEMU installed on the
Github runner, as newer versions of OVMF work fine in my local testing.