book: Set driver link-arg in `build.rs`
`config.toml` is a project-level configuration. Cargo will only read the
configuration from the current directory and up, and not in:

- package directories in a workspace
- package directories when `--manifest-path` is used

Use `build.rs` instead, which will be executed before the package is
built and only apply the flags for the specific package. This allows mixing
drivers and applications in a workspace while still invoking Cargo from the
workspace root.

Ref: https://doc.rust-lang.org/cargo/reference/config.html#hierarchical-structure
Ref: https://doc.rust-lang.org/cargo/reference/build-scripts.html
Signed-off-by: Tim Crawford <tcrawford@system76.com>