test-runner: Clean up device path tests
Split the conversion tests out into separate functions. This is a bit more
verbose, but makes it clearer exactly what is being tested.

The conversion tests now use a hardcoded test `DevicePath`, so we no longer need
to `cfg` the tests for different arches based on what devices QEMU creates. This
also makes it clearer what effects `DisplayOnly` and `AllowShortcuts` can have
on the text conversion.