xtask: Add file header check to xtask fmt
This will be used to verify that source files (just *.rs for now) all have an
SPDX license header. In non-check mode, the license header will be added.

For now, in check mode an error is shown but it's not treated as fatal. Once all
files have been fixed, the error will be made fatal.