doc: streamline
Streamline all `# Errors` and `# Arguments` sections. This is about
the style, not the content. The advantage of this streamlined and
consistent style, is that new contributions can follow the guideline.

This however is only a partial solution, in case we ever want to have a
100% streamlined documentation across the whole crate.

No major changes have been made with a few small exceptions, where I
deleted obviously wrong or irrelevant error codes.