uefi: allocator: use shortcut for PAGE_SIZE + add test
Allocating page-aligned memory via the global allocator is not
uncommon for UEFI OS loaders. Therefore, it is feasible to use a
shortcut in the allocator, and directly use boot::allocate_pages()
rather than boot::allocate_pool().

We can look at the TRACE messages of `cargo xtask run` to
verify that the shortcut is taken.