uefi/test-runner: Remove crate:: and uefi:: prefix from Status
In many places, especially doc comments, `Status` was referred to as
`uefi::Status` or `crate::Status`. Consistently use `Status` instead, adding
imports as needed. Note that in cases where `Status` is only used in docstrings,
the `use` is gated by `#[cfg(doc)]`. (This avoids needing a bunch of manual
`Status::*` links in docstrings.)

This is mostly a mechanical find-replace change, except for adding imports where
needed.