uefi: remove support for unstable allocator_api feature
The allocator_api feature [0] is old and not developed in years. Since
then, understanding of memory safety and best practises has evolved.
It is unlikely that in its current form the functionality will ever
be merged.

Therefore, we drop the complexity we have from this feature for now,
leading to simpler code.

[0] https://github.com/rust-lang/rust/issues/32838