uefi: serial: fix broken read handling
We didn't handle partial reads correctly. The serial protocol will
return Status::TIMEOUT in case it could not read as many bytes as
the caller asked for. This is however still the good case and just
a normal partial read. Therefore, I made the abstraction more
convenient and natural to use by handling this properly.